Adding the missing action examples, description. Issue: #6
