+++ plugins/quarkus-backend/src/scaffolder/actions/quarkus/createApp.example.ts
@@ -1,0 +1,28 @@
+import {TemplateExample} from '@backstage/plugin-scaffolder-node';
+import yaml from 'yaml';
+
+export const examples: TemplateExample[] = [
+    {
+        description: 'Clones a Quarkus quickstart project from GitHub',
+        example: yaml.stringify({
+            steps: [
+                {
+                    action: 'quarkus:app:create',
+                    id: 'quarkus-app-create',
+                    name: 'Create a Quarkus app',
+                    input: {
+                        values: {
+                            groupId: 'io.quarkus',
+                            artifactId: 'cool-demo',
+                            version: '1.0',
+                            javaVersion: '11',
+                            buildTool: 'MAVEN',
+                            quickstartName: 'quarkus-getting-started',
+                            extensions: ['quarkus-resteasy-reactive-jackson', 'quarkus-kubernetes', 'io.quarkus:quarkus-hibernate-orm-panache']
+                        },
+                    },
+                },
+            ],
+        }),
+    },
+];
