+++ plugins/quarkus-backend/src/scaffolder/actions/quarkus/index.ts
@@ -1,0 +1,2 @@
+export { createQuarkusApp } from './createApp';
+export { cloneQuarkusQuickstart } from './quickstartClone';
